requirements.txt
django==1.10.3
django-debug-toolbar==1.4
djangorestframework==3.3.3

mongoengine==0.10.6
pymongo==3.3.1

django-rest-framework-mongoengine==3.3.0